feat: Refactor into a component -> BackColorSelector
